Update README.md
In this repository I stated my name and email address associated with my github account.

along with stating the above information I gave a brief description of my github account reporting the number of repositories and specifically naming the current one.

Lastly, I noted the steps it took for me to clone this repository from github to my local device.